admin-app/requirements.txt
Flask==2.0.2
requests==2.25.1
urllib3==1.26.2
cryptography==3.4.8
oscrypto==1.2.1
pycryptodomex==3.10.1
asn1crypto==1.4.0
rsa==4.7
pyOpenSSL==21.0.0
PyJWT==2.3.0